+++ client/views/Team/TeamCard.jsx
@@ -1,3 +1,24 @@
+function addTeamMenuItem(menu,item,team) {
+	if(
+		team&&
+		team.hasSupplier(item)&&
+		team.canRemoveSupplier&&
+		team.canRemoveSupplier()&&
+		team._id!=item._id
+	) {
+
+		menu.push({
+			label:"Remove supplier from "+team.getName(),
+			shouldConfirm:true,
+			action(){
+				team.removeSupplier(item);
+				Modal.hide();
+			}
+		});
+
+	}
+}
+
 TeamCard = React.createClass({
 
 	getInitialState() {
@@ -15,40 +36,44 @@ TeamCard = React.createClass({
 	getMenu() {
 		var component = this;
 		var item = this.props.item;
-		var selectedTeam = Session.getSelectedTeam();
+		var parentTeam = Session.getSelectedTeam();
+		var parentFacility = Session.getSelectedFacility();
 		var menu = [];
-		
-		if(item&&item.canSave()) {
-			menu.push({
-				label:this.state.edit?"View as card":"Edit",
-				action(){
-					component.toggleEdit()
-				}
-			});
-		}
 
-		if(selectedTeam.canRemoveSupplier&&selectedTeam.canRemoveSupplier()&&selectedTeam._id!=item._id) {
-			menu.push({
-				label:"Remove supplier from your team",
-				shouldConfirm:true,
-				action(){
-					selectedTeam.removeSupplier(item);
-				}
-			});
+		if(item) {
+
+			if(item.canSave()) {
+				menu.push({
+					label:this.state.edit?"View as card":"Edit",
+					action(){
+						component.toggleEdit()
+					}
+				});
+			}
+
+			addTeamMenuItem(menu,item,parentTeam);
+			addTeamMenuItem(menu,item,parentFacility);
+
 		}
 		return menu;
 	},
 
 	render() {
 		var menu = this.getMenu();
-		var item = this.props.item;
+		var supplier = this.props.item;
 
 		return (
 			<div>
-			    {item.canSave()&&this.state.edit?
-			        <TeamViewEdit item={item} />
+			    {(!supplier||supplier.canSave())&&this.state.edit?
+			        <TeamViewEdit 
+						item={supplier} 
+						team={this.props.team}
+						facility={this.props.facility}
+						role={this.props.role}
+						onChange={this.props.onChange}
+			        />
 			    :
-					<TeamViewDetail item={item}/>
+					<TeamViewDetail item={supplier}/>
 				}
             	<ActionsMenu items={menu} />
 			</div>