+++ client/views/User/UserActions.jsx
@@ -2,50 +2,44 @@
 // actually it's kind of a controller - but then again so is the RBAC actions spec
 // can they be merged?
 
-UserActions = {
-	getMenu:function(user,context) {
-		var team = context.team;
-		var facility = context.facility;
+function addRoleBasedActions(menu,user,team) {
+	if(team&&team.hasMember(user)) {
+		var teamName = team.getName();
 
-		var menu = [];
-
-		if(team&&team.hasMember(user)) {
-
-			if(!user.isLoggedIn()) {
-				menu.push({
-					label:"Remove from "+team.getName(),
-					shouldConfirm:true,
-					action(){
-						team.removeMember(user);
-						Modal.hide();
-					}
-				});
-			}
-
-			if(team.canSetMemberRole&&team.canSetMemberRole(user)) {
-				var role = team.getRole(user);
-				var newRole = (role=='manager')?'staff':'manager';
-				menu.push({
-					label:(newRole=='manager')?'Promote to manager':'Demote to staff',
-					shouldConfirm:true,
-					action(){
-						team.setMemberRole(user,newRole);
-						Modal.hide();
-					}
-				})
-			}
-		}
-
-		if(facility&&facility.hasMember(user)) {
+		if(!user.isLoggedIn()) {
 			menu.push({
-				label:"Remove from "+facility.getName(),
+				label:"Remove from "+teamName,
 				shouldConfirm:true,
 				action(){
-					facility.removeMember(user);
+					team.removeMember(user);
 					Modal.hide();
 				}
 			});
 		}
+
+		if(team.canSetMemberRole&&team.canSetMemberRole(user)) {
+			var role = team.getMemberRole(user);
+			var newRole = (role=='manager')?'staff':'manager';
+			menu.push({
+				label:((newRole=='manager')?'Promote to manager':'Demote to staff')+(' of '+teamName),
+				shouldConfirm:true,
+				action(){
+					team.setMemberRole(user,newRole);
+					Modal.hide();
+				}
+			})
+		}
+	}
+}
+
+UserActions = {
+	getMenu:function(user,context) {
+		var team = context.team;
+		var facility = context.facility;
+		var menu = [];
+
+		addRoleBasedActions(menu,user,team);
+		addRoleBasedActions(menu,user,facility);
 
 		if(user) {
 			menu.push({